Added billable contract and renamed billable trait to HasSubscriptions
