+++ config/cashier.php
@@ -51,9 +51,9 @@ return [
     | Cashier Model
     |--------------------------------------------------------------------------
     |
-    | This is the model in your application that implements the Billable trait
-    | provided by Cashier. It will serve as the primary model you use while
-    | interacting with Cashier related methods, subscriptions, and so on.
+    | This is the model in your application that implements the Billable
+    | contract provided by Cashier. It will serve as the primary model you use
+    | while interacting with Cashier related methods, subscriptions, and so on.
     |
     */
