+++ src/Billable.php
@@ -2,24 +2,7 @@
 
 namespace Laravel\Cashier;
 
-use Exception;
-use Illuminate\Support\Collection;
-use Laravel\Cashier\Exceptions\InvalidStripeCustomer;
-use Stripe\BankAccount as StripeBankAccount;
-use Stripe\Card as StripeCard;
-use Stripe\Customer as StripeCustomer;
-use Stripe\Exception\CardException as StripeCardException;
-use Stripe\Exception\InvalidRequestException as StripeInvalidRequestException;
-use Stripe\Invoice as StripeInvoice;
-use Stripe\InvoiceItem as StripeInvoiceItem;
-use Stripe\PaymentIntent as StripePaymentIntent;
-use Stripe\PaymentMethod as StripePaymentMethod;
-use Stripe\Refund as StripeRefund;
-use Stripe\SetupIntent as StripeSetupIntent;
-use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
-
-trait Billable
+interface Billable
 {
     /**
      * Make a "one off" charge on the customer for the given amount.
@@ -29,29 +12,7 @@ trait Billable
      * @param  array  $options
      * @return \Laravel\Cashier\Payment
      */
-    public function charge($amount, $paymentMethod, array $options = [])
-    {
-        $options = array_merge([
-            'confirmation_method' => 'automatic',
-            'confirm' => true,
-            'currency' => $this->preferredCurrency(),
-        ], $options);
-
-        $options['amount'] = $amount;
-        $options['payment_method'] = $paymentMethod;
-
-        if ($this->stripe_id) {
-            $options['customer'] = $this->stripe_id;
-        }
-
-        $payment = new Payment(
-            StripePaymentIntent::create($options, $this->stripeOptions())
-        );
-
-        $payment->validate();
-
-        return $payment;
-    }
+    public function charge($amount, $paymentMethod, array $options = []);
 
     /**
      * Refund a customer for a charge.
@@ -60,15 +21,7 @@ trait Billable
      * @param  array  $options
      * @return \Stripe\Refund
      */
-    public function refund($paymentIntent, array $options = [])
-    {
-        $intent = StripePaymentIntent::retrieve($paymentIntent, $this->stripeOptions());
-
-        return StripeRefund::create(
-            ['charge' => $intent->charges->data[0]->id] + $options,
-            $this->stripeOptions()
-        );
-    }
+    public function refund($paymentIntent, array $options = []);
 
     /**
      * Add an invoice item to the customer's upcoming invoice.
@@ -78,19 +31,7 @@ trait Billable
      * @param  array  $options
      * @return \Stripe\InvoiceItem
      */
-    public function tab($description, $amount, array $options = [])
-    {
-        $this->assertCustomerExists();
-
-        $options = array_merge([
-            'customer' => $this->stripe_id,
-            'amount' => $amount,
-            'currency' => $this->preferredCurrency(),
-            'description' => $description,
-        ], $options);
-
-        return StripeInvoiceItem::create($options, $this->stripeOptions());
-    }
+    public function tab($description, $amount, array $options = []);
 
     /**
      * Invoice the customer for the given amount and generate an invoice immediately.
@@ -101,12 +42,7 @@ trait Billable
      * @param  array  $invoiceOptions
      * @return \Laravel\Cashier\Invoice|bool
      */
-    public function invoiceFor($description, $amount, array $tabOptions = [], array $invoiceOptions = [])
-    {
-        $this->tab($description, $amount, $tabOptions);
-
-        return $this->invoice($invoiceOptions);
-    }
+    public function invoiceFor($description, $amount, array $tabOptions = [], array $invoiceOptions = []);
 
     /**
      * Begin creating a new subscription.
@@ -115,10 +51,7 @@ trait Billable
      * @param  string  $plan
      * @return \Laravel\Cashier\SubscriptionBuilder
      */
-    public function newSubscription($subscription, $plan)
-    {
-        return new SubscriptionBuilder($this, $subscription, $plan);
-    }
+    public function newSubscription($subscription, $plan);
 
     /**
      * Determine if the Stripe model is on trial.
@@ -127,31 +60,14 @@ trait Billable
      * @param  string|null  $plan
      * @return bool
      */
-    public function onTrial($subscription = 'default', $plan = null)
-    {
-        if (func_num_args() === 0 && $this->onGenericTrial()) {
-            return true;
-        }
-
-        $subscription = $this->subscription($subscription);
-
-        if (is_null($plan)) {
-            return $subscription && $subscription->onTrial();
-        }
-
-        return $subscription && $subscription->onTrial() &&
-               $subscription->stripe_plan === $plan;
-    }
+    public function onTrial($subscription = 'default', $plan = null);
 
     /**
      * Determine if the Stripe model is on a "generic" trial at the model level.
      *
      * @return bool
      */
-    public function onGenericTrial()
-    {
-        return $this->trial_ends_at && $this->trial_ends_at->isFuture();
-    }
+    public function onGenericTrial();
 
     /**
      * Determine if the Stripe model has a given subscription.
@@ -160,21 +76,7 @@ trait Billable
      * @param  string|null  $plan
      * @return bool
      */
-    public function subscribed($subscription = 'default', $plan = null)
-    {
-        $subscription = $this->subscription($subscription);
-
-        if (is_null($subscription)) {
-            return false;
-        }
-
-        if (is_null($plan)) {
-            return $subscription->valid();
-        }
-
-        return $subscription->valid() &&
-               $subscription->stripe_plan === $plan;
-    }
+    public function subscribed($subscription = 'default', $plan = null);
 
     /**
      * Get a subscription instance by name.
@@ -182,24 +84,14 @@ trait Billable
      * @param  string  $subscription
      * @return \Laravel\Cashier\Subscription|null
      */
-    public function subscription($subscription = 'default')
-    {
-        return $this->subscriptions->sortByDesc(function ($value) {
-            return $value->created_at->getTimestamp();
-        })->first(function ($value) use ($subscription) {
-            return $value->name === $subscription;
-        });
-    }
+    public function subscription($subscription = 'default');
 
     /**
      * Get all of the subscriptions for the Stripe model.
      *
      * @return \Illuminate\Database\Eloquent\Relations\HasMany
      */
-    public function subscriptions()
-    {
-        return $this->hasMany(Subscription::class, $this->getForeignKey())->orderBy('created_at', 'desc');
-    }
+    public function subscriptions();
 
     /**
      * Determine if the customer's subscription has an incomplete payment.
@@ -207,14 +99,7 @@ trait Billable
      * @param  string  $subscription
      * @return bool
      */
-    public function hasIncompletePayment($subscription = 'default')
-    {
-        if ($subscription = $this->subscription($subscription)) {
-            return $subscription->hasIncompletePayment();
-        }
-
-        return false;
-    }
+    public function hasIncompletePayment($subscription = 'default');
 
     /**
      * Invoice the billable entity outside of the regular billing cycle.
@@ -222,50 +107,14 @@ trait Billable
      * @param  array  $options
      * @return \Laravel\Cashier\Invoice|bool
      */
-    public function invoice(array $options = [])
-    {
-        $this->assertCustomerExists();
-
-        $parameters = array_merge($options, ['customer' => $this->stripe_id]);
-
-        try {
-            /** @var \Stripe\Invoice $invoice */
-            $stripeInvoice = StripeInvoice::create($parameters, $this->stripeOptions());
-
-            $stripeInvoice = $stripeInvoice->pay();
-
-            return new Invoice($this, $stripeInvoice);
-        } catch (StripeInvalidRequestException $exception) {
-            return false;
-        } catch (StripeCardException $exception) {
-            $payment = new Payment(
-                StripePaymentIntent::retrieve(
-                    ['id' => $stripeInvoice->refresh()->payment_intent, 'expand' => ['invoice.subscription']],
-                    $this->stripeOptions()
-                )
-            );
-
-            $payment->validate();
-        }
-    }
+    public function invoice(array $options = []);
 
     /**
      * Get the entity's upcoming invoice.
      *
      * @return \Laravel\Cashier\Invoice|null
      */
-    public function upcomingInvoice()
-    {
-        $this->assertCustomerExists();
-
-        try {
-            $stripeInvoice = StripeInvoice::upcoming(['customer' => $this->stripe_id], $this->stripeOptions());
-
-            return new Invoice($this, $stripeInvoice);
-        } catch (StripeInvalidRequestException $exception) {
-            //
-        }
-    }
+    public function upcomingInvoice();
 
     /**
      * Find an invoice by ID.
@@ -273,20 +122,7 @@ trait Billable
      * @param  string  $id
      * @return \Laravel\Cashier\Invoice|null
      */
-    public function findInvoice($id)
-    {
-        $stripeInvoice = null;
-
-        try {
-            $stripeInvoice = StripeInvoice::retrieve(
-                $id, $this->stripeOptions()
-            );
-        } catch (Exception $exception) {
-            //
-        }
-
-        return $stripeInvoice ? new Invoice($this, $stripeInvoice) : null;
-    }
+    public function findInvoice($id);
 
     /**
      * Find an invoice or throw a 404 or 403 error.
@@ -294,20 +130,7 @@ trait Billable
      * @param  string  $id
      * @return \Laravel\Cashier\Invoice
      */
-    public function findInvoiceOrFail($id)
-    {
-        $invoice = $this->findInvoice($id);
-
-        if (is_null($invoice)) {
-            throw new NotFoundHttpException;
-        }
-
-        if ($invoice->customer !== $this->stripe_id) {
-            throw new AccessDeniedHttpException;
-        }
-
-        return $invoice;
-    }
+    public function findInvoiceOrFail($id);
 
     /**
      * Create an invoice download Response.
@@ -316,10 +139,7 @@ trait Billable
      * @param  array  $data
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function downloadInvoice($id, array $data)
-    {
-        return $this->findInvoiceOrFail($id)->download($data);
-    }
+    public function downloadInvoice($id, array $data);
 
     /**
      * Get a collection of the entity's invoices.
@@ -328,32 +148,7 @@ trait Billable
      * @param  array  $parameters
      * @return \Illuminate\Support\Collection
      */
-    public function invoices($includePending = false, $parameters = [])
-    {
-        $this->assertCustomerExists();
-
-        $invoices = [];
-
-        $parameters = array_merge(['limit' => 24], $parameters);
-
-        $stripeInvoices = StripeInvoice::all(
-            ['customer' => $this->stripe_id] + $parameters,
-            $this->stripeOptions()
-        );
-
-        // Here we will loop through the Stripe invoices and create our own custom Invoice
-        // instances that have more helper methods and are generally more convenient to
-        // work with than the plain Stripe objects are. Then, we'll return the array.
-        if (! is_null($stripeInvoices)) {
-            foreach ($stripeInvoices->data as $invoice) {
-                if ($invoice->paid || $includePending) {
-                    $invoices[] = new Invoice($this, $invoice);
-                }
-            }
-        }
-
-        return new Collection($invoices);
-    }
+    public function invoices($includePending = false, $parameters = []);
 
     /**
      * Get an array of the entity's invoices.
@@ -361,10 +156,7 @@ trait Billable
      * @param  array  $parameters
      * @return \Illuminate\Support\Collection
      */
-    public function invoicesIncludingPending(array $parameters = [])
-    {
-        return $this->invoices(true, $parameters);
-    }
+    public function invoicesIncludingPending(array $parameters = []);
 
     /**
      * Create a new SetupIntent instance.
@@ -372,32 +164,21 @@ trait Billable
      * @param  array  $options
      * @return \Stripe\SetupIntent
      */
-    public function createSetupIntent(array $options = [])
-    {
-        return StripeSetupIntent::create(
-            $options, $this->stripeOptions()
-        );
-    }
+    public function createSetupIntent(array $options = []);
 
     /**
      * Determines if the customer currently has a default payment method.
      *
      * @return bool
      */
-    public function hasDefaultPaymentMethod()
-    {
-        return (bool) $this->card_brand;
-    }
+    public function hasDefaultPaymentMethod();
 
     /**
      * Determines if the customer currently has at least one payment method.
      *
      * @return bool
      */
-    public function hasPaymentMethod()
-    {
-        return $this->paymentMethods()->isNotEmpty();
-    }
+    public function hasPaymentMethod();
 
     /**
      * Get a collection of the entity's payment methods.
@@ -405,22 +186,7 @@ trait Billable
      * @param  array  $parameters
      * @return \Illuminate\Support\Collection|\Laravel\Cashier\PaymentMethod[]
      */
-    public function paymentMethods($parameters = [])
-    {
-        $this->assertCustomerExists();
-
-        $parameters = array_merge(['limit' => 24], $parameters);
-
-        // "type" is temporarily required by Stripe...
-        $paymentMethods = StripePaymentMethod::all(
-            ['customer' => $this->stripe_id, 'type' => 'card'] + $parameters,
-            $this->stripeOptions()
-        );
-
-        return collect($paymentMethods->data)->map(function ($paymentMethod) {
-            return new PaymentMethod($this, $paymentMethod);
-        });
-    }
+    public function paymentMethods($parameters = []);
 
     /**
      * Add a payment method to the customer.
@@ -428,20 +194,7 @@ trait Billable
      * @param  \Stripe\PaymentMethod|string  $paymentMethod
      * @return \Laravel\Cashier\PaymentMethod
      */
-    public function addPaymentMethod($paymentMethod)
-    {
-        $this->assertCustomerExists();
-
-        $stripePaymentMethod = $this->resolveStripePaymentMethod($paymentMethod);
-
-        if ($stripePaymentMethod->customer !== $this->stripe_id) {
-            $stripePaymentMethod = $stripePaymentMethod->attach(
-                ['customer' => $this->stripe_id], $this->stripeOptions()
-            );
-        }
-
-        return new PaymentMethod($this, $stripePaymentMethod);
-    }
+    public function addPaymentMethod($paymentMethod);
 
     /**
      * Remove a payment method from the customer.
@@ -449,57 +202,14 @@ trait Billable
      * @param  \Stripe\PaymentMethod|string  $paymentMethod
      * @return void
      */
-    public function removePaymentMethod($paymentMethod)
-    {
-        $this->assertCustomerExists();
-
-        $stripePaymentMethod = $this->resolveStripePaymentMethod($paymentMethod);
-
-        if ($stripePaymentMethod->customer !== $this->stripe_id) {
-            return;
-        }
-
-        $customer = $this->asStripeCustomer();
-
-        $defaultPaymentMethod = $customer->invoice_settings->default_payment_method;
-
-        $stripePaymentMethod->detach(null, $this->stripeOptions());
-
-        // If the payment method was the default payment method, we'll remove it manually...
-        if ($stripePaymentMethod->id === $defaultPaymentMethod) {
-            $this->forceFill([
-                'card_brand' => null,
-                'card_last_four' => null,
-            ])->save();
-        }
-    }
+    public function removePaymentMethod($paymentMethod);
 
     /**
      * Get the default payment method for the entity.
      *
      * @return \Laravel\Cashier\PaymentMethod|\Stripe\Card|\Stripe\BankAccount|null
      */
-    public function defaultPaymentMethod()
-    {
-        if (! $this->hasStripeId()) {
-            return;
-        }
-
-        $customer = StripeCustomer::retrieve([
-            'id' => $this->stripe_id,
-            'expand' => [
-                'invoice_settings.default_payment_method',
-                'default_source',
-            ],
-        ], $this->stripeOptions());
-
-        if ($customer->invoice_settings->default_payment_method) {
-            return new PaymentMethod($this, $customer->invoice_settings->default_payment_method);
-        }
-
-        // If we can't find a payment method, try to return a legacy source...
-        return $customer->default_source;
-    }
+    public function defaultPaymentMethod();
 
     /**
      * Update customer's default payment method.
@@ -507,114 +217,21 @@ trait Billable
      * @param  \Stripe\PaymentMethod|string  $paymentMethod
      * @return \Laravel\Cashier\PaymentMethod
      */
-    public function updateDefaultPaymentMethod($paymentMethod)
-    {
-        $this->assertCustomerExists();
-
-        $customer = $this->asStripeCustomer();
-
-        $stripePaymentMethod = $this->resolveStripePaymentMethod($paymentMethod);
-
-        // If the customer already has the payment method as their default, we can bail out
-        // of the call now. We don't need to keep adding the same payment method to this
-        // model's account every single time we go through this specific process call.
-        if ($stripePaymentMethod->id === $customer->invoice_settings->default_payment_method) {
-            return;
-        }
-
-        $paymentMethod = $this->addPaymentMethod($stripePaymentMethod);
-
-        $customer->invoice_settings = ['default_payment_method' => $paymentMethod->id];
-
-        $customer->save($this->stripeOptions());
-
-        // Next we will get the default payment method for this user so we can update the
-        // payment method details on the record in the database. This will allow us to
-        // show that information on the front-end when updating the payment methods.
-        $this->fillPaymentMethodDetails($paymentMethod);
-
-        $this->save();
-
-        return $paymentMethod;
-    }
+    public function updateDefaultPaymentMethod($paymentMethod);
 
     /**
      * Synchronises the customer's default payment method from Stripe back into the database.
      *
      * @return $this
      */
-    public function updateDefaultPaymentMethodFromStripe()
-    {
-        $defaultPaymentMethod = $this->defaultPaymentMethod();
-
-        if ($defaultPaymentMethod) {
-            if ($defaultPaymentMethod instanceof PaymentMethod) {
-                $this->fillPaymentMethodDetails(
-                    $defaultPaymentMethod->asStripePaymentMethod()
-                )->save();
-            } else {
-                $this->fillSourceDetails($defaultPaymentMethod)->save();
-            }
-        } else {
-            $this->forceFill([
-                'card_brand' => null,
-                'card_last_four' => null,
-            ])->save();
-        }
-
-        return $this;
-    }
-
-    /**
-     * Fills the model's properties with the payment method from Stripe.
-     *
-     * @param  \Laravel\Cashier\PaymentMethod|\Stripe\PaymentMethod|null  $paymentMethod
-     * @return $this
-     */
-    protected function fillPaymentMethodDetails($paymentMethod)
-    {
-        if ($paymentMethod->type === 'card') {
-            $this->card_brand = $paymentMethod->card->brand;
-            $this->card_last_four = $paymentMethod->card->last4;
-        }
-
-        return $this;
-    }
-
-    /**
-     * Fills the model's properties with the source from Stripe.
-     *
-     * @param  \Stripe\Card|\Stripe\BankAccount|null  $source
-     * @return $this
-     *
-     * @deprecated Will be removed in a future Cashier update. You should use the new payment methods API instead.
-     */
-    protected function fillSourceDetails($source)
-    {
-        if ($source instanceof StripeCard) {
-            $this->card_brand = $source->brand;
-            $this->card_last_four = $source->last4;
-        } elseif ($source instanceof StripeBankAccount) {
-            $this->card_brand = 'Bank Account';
-            $this->card_last_four = $source->last4;
-        }
-
-        return $this;
-    }
+    public function updateDefaultPaymentMethodFromStripe();
 
     /**
      * Deletes the entity's payment methods.
      *
      * @return void
      */
-    public function deletePaymentMethods()
-    {
-        $this->paymentMethods()->each(function (PaymentMethod $paymentMethod) {
-            $paymentMethod->delete();
-        });
-
-        $this->updateDefaultPaymentMethodFromStripe();
-    }
+    public function deletePaymentMethods();
 
     /**
      * Find a PaymentMethod by ID.
@@ -622,35 +239,7 @@ trait Billable
      * @param  string  $paymentMethod
      * @return \Laravel\Cashier\PaymentMethod|null
      */
-    public function findPaymentMethod($paymentMethod)
-    {
-        $stripePaymentMethod = null;
-
-        try {
-            $stripePaymentMethod = $this->resolveStripePaymentMethod($paymentMethod);
-        } catch (Exception $exception) {
-            //
-        }
-
-        return $stripePaymentMethod ? new PaymentMethod($this, $stripePaymentMethod) : null;
-    }
-
-    /**
-     * Resolve a PaymentMethod ID to a Stripe PaymentMethod object.
-     *
-     * @param  \Stripe\PaymentMethod|string  $paymentMethod
-     * @return \Stripe\PaymentMethod
-     */
-    protected function resolveStripePaymentMethod($paymentMethod)
-    {
-        if ($paymentMethod instanceof StripePaymentMethod) {
-            return $paymentMethod;
-        }
-
-        return StripePaymentMethod::retrieve(
-            $paymentMethod, $this->stripeOptions()
-        );
-    }
+    public function findPaymentMethod($paymentMethod);
 
     /**
      * Apply a coupon to the billable entity.
@@ -658,16 +247,7 @@ trait Billable
      * @param  string  $coupon
      * @return void
      */
-    public function applyCoupon($coupon)
-    {
-        $this->assertCustomerExists();
-
-        $customer = $this->asStripeCustomer();
-
-        $customer->coupon = $coupon;
-
-        $customer->save();
-    }
+    public function applyCoupon($coupon);
 
     /**
      * Determine if the Stripe model is actively subscribed to one of the given plans.
@@ -676,22 +256,7 @@ trait Billable
      * @param  string  $subscription
      * @return bool
      */
-    public function subscribedToPlan($plans, $subscription = 'default')
-    {
-        $subscription = $this->subscription($subscription);
-
-        if (! $subscription || ! $subscription->valid()) {
-            return false;
-        }
-
-        foreach ((array) $plans as $plan) {
-            if ($subscription->stripe_plan === $plan) {
-                return true;
-            }
-        }
-
-        return false;
-    }
+    public function subscribedToPlan($plans, $subscription = 'default');
 
     /**
      * Determine if the entity is on the given plan.
@@ -699,36 +264,14 @@ trait Billable
      * @param  string  $plan
      * @return bool
      */
-    public function onPlan($plan)
-    {
-        return ! is_null($this->subscriptions->first(function ($value) use ($plan) {
-            return $value->stripe_plan === $plan && $value->valid();
-        }));
-    }
+    public function onPlan($plan);
 
     /**
      * Determine if the entity has a Stripe customer ID.
      *
      * @return bool
      */
-    public function hasStripeId()
-    {
-        return ! is_null($this->stripe_id);
-    }
-
-    /**
-     * Determine if the entity has a Stripe customer ID and throw an exception if not.
-     *
-     * @return void
-     *
-     * @throws \Laravel\Cashier\Exceptions\InvalidStripeCustomer
-     */
-    protected function assertCustomerExists()
-    {
-        if (! $this->stripe_id) {
-            throw InvalidStripeCustomer::nonCustomer($this);
-        }
-    }
+    public function hasStripeId();
 
     /**
      * Create a Stripe customer for the given model.
@@ -736,25 +279,7 @@ trait Billable
      * @param  array  $options
      * @return \Stripe\Customer
      */
-    public function createAsStripeCustomer(array $options = [])
-    {
-        $options = array_key_exists('email', $options)
-                ? $options
-                : array_merge($options, ['email' => $this->email]);
-
-        // Here we will create the customer instance on Stripe and store the ID of the
-        // user from Stripe. This ID will correspond with the Stripe user instances
-        // and allow us to retrieve users from Stripe later when we need to work.
-        $customer = StripeCustomer::create(
-            $options, $this->stripeOptions()
-        );
-
-        $this->stripe_id = $customer->id;
-
-        $this->save();
-
-        return $customer;
-    }
+    public function createAsStripeCustomer(array $options = []);
 
     /**
      * Update the underlying Stripe customer information for the model.
@@ -762,12 +287,7 @@ trait Billable
      * @param  array  $options
      * @return \Stripe\Customer
      */
-    public function updateStripeCustomer(array $options = [])
-    {
-        return StripeCustomer::update(
-            $this->stripe_id, $options, $this->stripeOptions()
-        );
-    }
+    public function updateStripeCustomer(array $options = []);
 
     /**
      * Get the Stripe customer instance for the current user or create one.
@@ -775,46 +295,28 @@ trait Billable
      * @param  array  $options
      * @return \Stripe\Customer
      */
-    public function createOrGetStripeCustomer(array $options = [])
-    {
-        if ($this->stripe_id) {
-            return $this->asStripeCustomer();
-        }
-
-        return $this->createAsStripeCustomer($options);
-    }
+    public function createOrGetStripeCustomer(array $options = []);
 
     /**
      * Get the Stripe customer for the model.
      *
      * @return \Stripe\Customer
      */
-    public function asStripeCustomer()
-    {
-        $this->assertCustomerExists();
-
-        return StripeCustomer::retrieve($this->stripe_id, $this->stripeOptions());
-    }
+    public function asStripeCustomer();
 
     /**
      * Get the Stripe supported currency used by the entity.
      *
      * @return string
      */
-    public function preferredCurrency()
-    {
-        return config('cashier.currency');
-    }
+    public function preferredCurrency();
 
     /**
      * Get the tax percentage to apply to the subscription.
      *
      * @return int|float
      */
-    public function taxPercentage()
-    {
-        return 0;
-    }
+    public function taxPercentage();
 
     /**
      * Get the default Stripe API options for the current Billable model.
@@ -822,8 +324,5 @@ trait Billable
      * @param  array  $options
      * @return array
      */
-    public function stripeOptions(array $options = [])
-    {
-        return Cashier::stripeOptions($options);
-    }
+    public function stripeOptions(array $options = []);
 }
